c2-il: IL tokens are variable width (2 or 4 bytes), not a per-file constant
detect_token_width returns one width for a whole .ex, which is wrong for
real translation units: a single file carries both. Captured from
system/world/Dir.cpp, the same 4F 02 module marker appears as

    4f 02 e3 09            (2-byte token)
    4f 02 a4 96 03 00      (4-byte token)

The discriminator is bit 7 of the token's SECOND byte: clear = 2 bytes,
set = 2 more follow. Applying the rule to every B9 LOAD site in that TU
lands on a valid 3-byte operand type at 21443 sites, the residue being a
third type class plus B9 bytes occurring inside data.

A 2-byte read of a 4-byte token leaves the parse standing on the token's
own tail bytes, which then look like unknown opcodes. That was fabricating
whole families of census buckets: call-token-0x01..0x05 and
expr-load-type-0N00A6 were misalignment, not vocabulary. Both families
disappear with this fix.

Token identity widens u16 -> u32; the two widths cannot alias (a 4-byte
token's bit 7 lands in bits 23..16, forcing the value >= 0x10000, while
2-byte values stay below it). detect_token_width is kept for the K1/K2a
codec and its tests but the parser no longer consults it. NOTE: the codec
still assumes a fixed 2-byte token (codec::tok16) and has the same latent
issue on real TUs — tracked separately, it is round-trip gated so it fails
closed rather than mis-decoding.

Gate unchanged: 25/25 fixtures replay byte-exact, 13/25 Port=Match,
0 mismatch, workspace tests green.
Workload census: 4154 -> 7114 / 2462571 functions in class (0.17% -> 0.29%).

+++ crates/c2-core/src/codegen.rs
@@ -321,7 +321,7 @@ pub fn select_text(func: &IlFunction) -> Result<Vec<u8>, BackendError> {
     }
 
     // token -> incoming ABI register, by declaration order.
-    let reg_of = |tok: u16| -> Option<u8> {
+    let reg_of = |tok: u32| -> Option<u8> {
         func.params
             .iter()
             .position(|&t| t == tok)
@@ -627,7 +627,7 @@ mod tests {
         assert_eq!(encode_addi(3, 0, 42), [0x38, 0x60, 0x00, 0x2A]); // li r3,42
     }
 
-    fn func_with(params: Vec<u16>, ops: Vec<IlOp>) -> IlFunction {
+    fn func_with(params: Vec<u32>, ops: Vec<IlOp>) -> IlFunction {
         IlFunction {
             mangled_name: "?f@@YAHH@Z".into(),
             source_path: None,
